+++ crates/config/src/toolchain_config.rs
@@ -6,7 +6,7 @@ use schematic::{Config, validate};
 use version_spec::UnresolvedVersionSpec;
 
 #[cfg(feature = "proto")]
-use crate::{inherit_tool, inherit_tool_without_version, is_using_tool_version};
+use crate::{inherit_tool, is_using_tool_version};
 
 /// Configures all tools and platforms.
 /// Docs: https://moonrepo.dev/docs/config/toolchain
@@ -47,10 +47,6 @@ pub struct ToolchainConfig {
     /// Configures and enables the Rust platform.
     #[setting(nested)]
     pub rust: Option<RustConfig>,
-
-    /// Configures and enables the TypeScript platform.
-    #[setting(nested)]
-    pub typescript: Option<TypeScriptConfig>,
 
     /// All configured toolchains by unique ID.
     #[setting(flatten, nested)]
@@ -194,13 +190,6 @@ impl ToolchainConfig {
 
     inherit_tool!(RustConfig, rust, "rust", inherit_proto_rust);
 
-    inherit_tool_without_version!(
-        TypeScriptConfig,
-        typescript,
-        "typescript",
-        inherit_proto_typescript
-    );
-
     pub fn should_install_proto(&self) -> bool {
         is_using_tool_version!(self, bun);
         is_using_tool_version!(self, deno);
@@ -230,7 +219,6 @@ impl ToolchainConfig {
         self.inherit_proto_node(proto_config)?;
         self.inherit_proto_python(proto_config)?;
         self.inherit_proto_rust(proto_config)?;
-        self.inherit_proto_typescript(proto_config)?;
 
         if let Some(node_config) = &mut self.node {
             node_config.inherit_proto(proto_config)?;